refactor: move option machinery out of the model file
nox/_options.py now holds only the declarative option model. The shell
completers move to nox/_completers.py (dropping the nox.tasks import from
the model), and the post-parse policy (alias finalization and the noxfile
merge) moves to nox/_merge.py.
