+++ nox/_options.py
@@ -16,32 +16,21 @@
 
 from __future__ import annotations
 
-__lazy_modules__ = {
-    "argcomplete",
-    "argparse",
-    "itertools",
-    "nox._option_set",
-    "nox.tasks",
-    "nox.virtualenv",
-}
+__lazy_modules__ = {"argparse", "nox._option_set", "nox.virtualenv"}
 
 import argparse
-import itertools
 import os
-import sys
 from typing import TYPE_CHECKING, Any, Literal
 
-import argcomplete
 import attrs
 import attrs.validators as av
 
-from nox import _option_set
-from nox._option_set import Forward, Source, opt
-from nox.tasks import discover_manifest, filter_manifest, load_nox_module
+from nox import _completers, _merge, _option_set
+from nox._option_set import Forward, opt
 from nox.virtualenv import ALL_VENVS
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterable, Sequence
+    from collections.abc import Callable
 
     # Opaque validator type; keeps mypy's attrs plugin from inferring field
     # types from the validator instead of the annotation.
@@ -52,7 +41,6 @@ __all__ = [
     "NoxConfig",
     "NoxfileOptions",
     "ReuseVenvType",
-    "merge_noxfile_options",
     "noxfile_options",
     "options",
 ]
@@ -130,53 +118,6 @@ def _serialize_color(value: Any) -> list[str]:
     return ["--forcecolor"] if value else ["--nocolor"]
 
 
-def _python_completer(
-    prefix: str,  # noqa: ARG001
-    parsed_args: argparse.Namespace | NoxConfig,
-    **kwargs: Any,
-) -> Iterable[str]:
-    config = options.expand(parsed_args)
-    module = load_nox_module(config)
-    manifest = discover_manifest(module, config)
-    return filter(
-        None,
-        (
-            session.func.python  # type:ignore[misc] # str sequences flattened, other non-strs falsey and filtered out
-            for session, _ in manifest.list_all_sessions()
-        ),
-    )
-
-
-def _session_completer(
-    prefix: str,  # noqa: ARG001
-    parsed_args: argparse.Namespace | NoxConfig,
-    **kwargs: Any,
-) -> Iterable[str]:
-    config = options.expand(parsed_args)
-    config.list_sessions = True
-    module = load_nox_module(config)
-    manifest = discover_manifest(module, config)
-    filtered_manifest = filter_manifest(manifest, config)
-    if isinstance(filtered_manifest, int):
-        return []
-    return (
-        session.friendly_name for session, _ in filtered_manifest.list_all_sessions()
-    )
-
-
-def _tag_completer(
-    prefix: str,  # noqa: ARG001
-    parsed_args: argparse.Namespace | NoxConfig,
-    **kwargs: Any,
-) -> Iterable[str]:
-    config = options.expand(parsed_args)
-    module = load_nox_module(config)
-    manifest = discover_manifest(module, config)
-    return itertools.chain.from_iterable(
-        filter(None, (session.tags for session, _ in manifest.list_all_sessions()))
-    )
-
-
 @attrs.define(
     kw_only=True,
     on_setattr=[attrs.setters.validate, _option_set.record_noxfile_set],
@@ -226,7 +167,7 @@ class NoxfileOptions(_option_set.OptionsBase):
         metadata=opt(
             "--envdir",
             group="environment",
-            completer=argcomplete.completers.DirectoriesCompleter(),  # type: ignore[no-untyped-call]
+            completer=_completers.directory_completer,
             help="Directory where Nox will store virtualenvs, this is ``.nox`` by default.",
         ),
     )
@@ -277,7 +218,7 @@ class NoxfileOptions(_option_set.OptionsBase):
             "-k",
             "--keywords",
             group="sessions",
-            completer=argcomplete.completers.ChoicesCompleter(()),  # type: ignore[no-untyped-call]
+            completer=_completers.empty_completer,
             help="Only run sessions that match the given expression.",
         ),
     )
@@ -290,7 +231,7 @@ class NoxfileOptions(_option_set.OptionsBase):
             "--python",
             group="python",
             env_var="NOXPYTHON",
-            completer=_python_completer,
+            completer=_completers.python_completer,
             help=(
                 "Only run sessions that use the given python interpreter versions."
                 " Environment variable: NOXPYTHON"
@@ -303,7 +244,7 @@ class NoxfileOptions(_option_set.OptionsBase):
         metadata=opt(
             "--report",
             group="reporting",
-            completer=argcomplete.completers.FilesCompleter(("json",)),  # type: ignore[no-untyped-call]
+            completer=_completers.json_file_completer,
             help="Output a report of all sessions to the given filename.",
         ),
     )
@@ -341,7 +282,7 @@ class NoxfileOptions(_option_set.OptionsBase):
             "--session",
             group="sessions",
             env_var="NOXSESSION",
-            completer=_session_completer,
+            completer=_completers.session_completer,
             help=(
                 "Which sessions to run. By default, all sessions will run."
                 " Environment variable: NOXSESSION"
@@ -367,7 +308,7 @@ class NoxfileOptions(_option_set.OptionsBase):
             "-t",
             "--tags",
             group="sessions",
-            completer=_tag_completer,
+            completer=_completers.tag_completer,
             help="Only run sessions with the given tags.",
         ),
     )
@@ -477,7 +418,7 @@ class NoxConfig(NoxfileOptions):
             "--extra-python",
             group="python",
             env_var="NOXEXTRAPYTHON",
-            completer=_python_completer,
+            completer=_completers.python_completer,
             help=(
                 "Additionally, run sessions using the given python interpreter versions."
                 " Environment variable: NOXEXTRAPYTHON"
@@ -492,7 +433,7 @@ class NoxConfig(NoxfileOptions):
             "--force-python",
             group="python",
             env_var="NOXFORCEPYTHON",
-            completer=_python_completer,
+            completer=_completers.python_completer,
             help=(
                 "Run sessions with the given interpreters instead of those listed in the"
                 " Noxfile. This is a shorthand for ``--python=X.Y --extra-python=X.Y``."
@@ -598,93 +539,11 @@ class NoxConfig(NoxfileOptions):
     )
 
 
-def _strip_posargs(posargs: Sequence[str]) -> list[str]:
-    """Remove the leading "--" from the posargs array (if any) and assert that
-    the remaining arguments came after a "--"."""
-    if not posargs:
-        return []
-
-    dash_index = posargs.index("--") if "--" in posargs else len(posargs)
-    if dash_index != 0:
-        unexpected_posargs = posargs[:dash_index]
-        raise _option_set.ArgumentError(
-            None, f"Unknown argument(s) '{' '.join(unexpected_posargs)}'."
-        )
-
-    return list(posargs[dash_index + 1 :])
-
-
-def _finalize(config: NoxConfig) -> None:
-    """Resolve aliases and cross-field options after parsing."""
-    config.posargs = _strip_posargs(config.posargs)
-
-    if config.force_pythons:
-        source = config.provenance("force_pythons")
-        config.set_value("pythons", config.force_pythons, source)
-        config.set_value("extra_pythons", config.force_pythons, source)
-
-    if config.R:
-        # -R wins even over an explicit --reuse-venv (long-standing behavior).
-        config.set_value("reuse_venv", "yes", Source.COMMAND_LINE)
-        config.set_value("reuse_existing_virtualenvs", True, Source.COMMAND_LINE)  # noqa: FBT003
-        config.set_value("no_install", True, Source.COMMAND_LINE)  # noqa: FBT003
-
-    # -r/-N are an alias for --reuse-venv=yes|no; an explicit --reuse-venv wins.
-    if (
-        config.reuse_existing_virtualenvs is not None
-        and config.provenance("reuse_venv") is not Source.COMMAND_LINE
-    ):
-        config.set_value(
-            "reuse_venv",
-            "yes" if config.reuse_existing_virtualenvs else "no",
-            Source.COMMAND_LINE,
-        )
-
-    if config.no_venv:
-        if config.force_venv_backend not in {None, "none"}:
-            msg = "You can not use `--no-venv` with a non-none `--force-venv-backend`"
-            raise ValueError(msg)
-        config.set_value("force_venv_backend", "none", Source.COMMAND_LINE)
-
-    if config.forcecolor and config.nocolor:
-        raise _option_set.ArgumentError(
-            None, "Can not specify both --no-color and --force-color."
-        )
-    config.color = config.forcecolor or (not config.nocolor and sys.stdout.isatty())
-
-
-def merge_noxfile_options(config: NoxConfig, noxfile_config: NoxfileOptions) -> None:
-    """Apply ``nox.options`` (as set by the noxfile) to the parsed config.
-
-    A value explicitly given on the command line (or via an environment
-    variable) always wins over the noxfile.
-    """
-    # sessions/keywords/tags are only taken from the noxfile if none of the
-    # three was given on the command line.
-    trio = ("sessions", "keywords", "tags")
-    cli_selected = any(config.provenance(name) is not Source.DEFAULT for name in trio)
-    _option_set.apply_noxfile_values(
-        config, noxfile_config, skip=trio if cli_selected else ()
-    )
-
-    # The legacy reuse_existing_virtualenvs alias, when enabled in the
-    # noxfile, wins over a noxfile-set reuse_venv (but not the command line).
-    if (
-        noxfile_config.provenance("reuse_existing_virtualenvs") is not Source.DEFAULT
-        and noxfile_config.reuse_existing_virtualenvs
-        and config.provenance("reuse_venv") is not Source.COMMAND_LINE
-    ):
-        config.set_value("reuse_venv", "yes", Source.NOXFILE)
-
-    # The noxfile may have set a PathLike envdir.
-    config.envdir = os.fspath(config.envdir)
-
-
 options = _option_set.Options(
     NoxConfig,
     groups=GROUPS,
     description="Nox is a Python automation toolkit.",
-    finalize=_finalize,
+    finalize=_merge.finalize,
 )
 
 noxfile_options: NoxfileOptions = NoxfileOptions()
